update union and message decoding to limit their read size instead of prereading whole records
move top level length logic to settings structure
